pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sleepless-agent"
version = "1.0.0"
description = "Minimal Claude Code continuous execution supervisor runtime"
readme = "README.md"
authors = [{name = "Sleepless Agent Maintainers"}]
license = {file = "LICENSE"}
keywords = ["ai", "automation", "agents", "claude", "supervisor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
]

[project.scripts]
sle = "sleepless_agent.cli:main"

[project.urls]
Homepage = "https://github.com/context-machine-lab/sleepless-agent"
Repository = "https://github.com/context-machine-lab/sleepless-agent"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sleepless_agent*"]
